requirements.txt
pyautogui>=0.9.54
pygetwindow>=0.0.9
Pillow>=9.0.0
opencv-python>=4.5.0
mss>=6.1.0
numpy>=1.21.0
pytesseract>=0.3.10
pywin32